src/main/component_3/requirements.txt
blinker==1.7.0
click==8.1.7
coverage==7.3.2
exceptiongroup==1.1.3
Flask==3.0.0
grpcio==1.60.0
grpcio-testing==1.60.0
grpcio-tools==1.59.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
protobuf==4.25.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
tomli==2.0.1
Werkzeug==3.0.1
